+++ src/nymphes_osc/NymphesOSC.py
@@ -481,11 +481,19 @@ class NymphesOSC:
 
     def _send_status_to_osc_clients(self, message):
         """
-        Sends a string status message to OSC clients, using the address /status.
+        Sends a status message string to OSC clients, using the address /status.
         Also logs the message.
         """
         # Send to all clients
         self._send_osc_to_all_clients('/status', str(message))
+
+    def _send_error_message_to_osc_clients(self, message, detailed_message):
+        """
+        Sends an error message string to OSC clients, using the address /error.
+        Also logs the message.
+        """
+        # Send to all clients
+        self._send_osc_to_all_clients('/error', str(message), str(detailed_message))
 
     def _send_osc_to_all_clients(self, address, *args):
         """
@@ -532,9 +540,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to register client ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to register client'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_register_client_with_ip_address(self, sender_ip, address, *args):
         """
@@ -560,9 +568,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to register client ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to register client'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_unregister_client(self, sender_ip, address, *args):
         """
@@ -587,9 +595,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to unregister client ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to unregister client'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_unregister_client_with_ip_address(self, sender_ip, address, *args):
         """
@@ -616,9 +624,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to unregister client ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to unregister client'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_load_preset(self, sender_ip, address, *args):
         """
@@ -648,9 +656,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to load preset ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to load preset'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_load_file(self, sender_ip, address, *args):
         """
@@ -666,18 +674,27 @@ class NymphesOSC:
             return
 
         try:
-            filepath = args[0]
-
+            filepath = Path(args[0])
             self.logger.info(f'Received {address} {filepath} from {sender_ip[0]}')
 
-            # Load the file
-            self._nymphes_midi.load_file(filepath=filepath)
+            if filepath.suffix == '.txt':
+                # This might be a preset file
+                self._nymphes_midi.load_file(filepath=filepath)
+
+            elif filepath.suffix == '.syx':
+                # This may be a sysex file containing a preset
+                self._nymphes_midi.load_syx_file(filepath=filepath)
+
+            else:
+                status = f'Invalid file: {filepath}'
+                self._send_error_message_to_osc_clients(status, '')
+                self.logger.warning(status)
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to load file into current preset({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to load file into current preset: {filepath}'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_save_to_preset(self, sender_ip, address, *args):
         """
@@ -709,9 +726,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to load current preset into Nymphes preset slot ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to load current preset into Nymphes preset slot'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_load_file_to_preset(self, sender_ip, address, *args):
         """
@@ -745,9 +762,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to load file into Nymphes preset slot ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to load file into Nymphes preset slot'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_save_to_file(self, sender_ip, address, *args):
         """
@@ -773,9 +790,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to save current preset to file ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to save current preset to file'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_save_preset_to_file(self, sender_ip, address, *args):
         """
@@ -809,9 +826,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to save preset from Nymphes preset slot to file ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to save preset from Nymphes preset slot to file'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_load_init_file(self, sender_ip, address, *args):
         """
@@ -827,9 +844,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to load init preset ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to load init preset'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_request_preset_dump(self, sender_ip, address, *args):
         """
@@ -867,9 +884,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to connect MIDI Input port ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to connect MIDI Input port'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_connect_nymphes(self, sender_ip, address, *args):
         """
@@ -898,9 +915,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to connect Nymphes ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to connect Nymphes'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_disconnect_nymphes(self, sender_ip, address, *args):
         """
@@ -918,9 +935,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to disconnect Nymphes ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to disconnect Nymphes'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_disconnect_midi_input(self, sender_ip, address, *args):
         """
@@ -945,9 +962,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to disconnect MIDI Input port ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to disconnect MIDI Input port'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_connect_midi_output(self, sender_ip, address, *args):
         """
@@ -972,9 +989,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to connect MIDI Output port ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to connect MIDI Output port'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_disconnect_midi_output(self, sender_ip, address, *args):
         """
@@ -999,9 +1016,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to disconnect MIDI Output port ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to disconnect MIDI Output port'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_set_nymphes_midi_channel(self, sender_ip, address, *args):
         """
@@ -1027,9 +1044,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to set Nymphes MIDI channel ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            message = f'Failed to set Nymphes MIDI channel'
+            self._send_error_message_to_osc_clients(message, str(e))
+            self.logger.warning(f'{message}: {e}')
 
     def _on_osc_message_mod_wheel(self, sender_ip, address, *args):
         """
@@ -1053,9 +1070,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to set mod wheel ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to set mod wheel'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_aftertouch(self, sender_ip, address, *args):
         """
@@ -1079,9 +1096,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to set aftertouch ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to set aftertouch'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_osc_message_sustain_pedal(self, sender_ip, address, *args):
         """
@@ -1106,9 +1123,9 @@ class NymphesOSC:
 
         except Exception as e:
             # Send status update and log it
-            status = f'Failed to set sustain_pedal ({e})'
-            self._send_status_to_osc_clients(status)
-            self.logger.warning(status)
+            status = f'Failed to set sustain_pedal'
+            self._send_error_message_to_osc_clients(status, str(e))
+            self.logger.warning(f'{status}: {e}')
 
     def _on_other_osc_message(self, sender_ip, address, *args):
         """
@@ -1146,9 +1163,9 @@ class NymphesOSC:
 
                 except Exception as e:
                     # Send status update and log it
-                    status = f'Failed to set parameter ({e})'
-                    self._send_status_to_osc_clients(status)
-                    self.logger.warning(status)
+                    status = f'Failed to set parameter'
+                    self._send_error_message_to_osc_clients(status, str(e))
+                    self.logger.warning(f'{status}: {e}')
 
             elif isinstance(value, float):
                 try:
@@ -1156,14 +1173,14 @@ class NymphesOSC:
 
                 except Exception as e:
                     # Send status update and log it
-                    status = f'Failed to set parameter ({e})'
-                    self._send_status_to_osc_clients(status)
-                    self.logger.warning(status)
+                    status = f'Failed to set parameter'
+                    self._send_error_message_to_osc_clients(status, str(e))
+                    self.logger.warning(f'{status}: {e}')
 
             else:
                 # Send status update and log it
                 status = f'Invalid value type for {param_name}: {type(value)}'
-                self._send_status_to_osc_clients(status)
+                self._send_error_message_to_osc_clients(status, '')
                 self.logger.warning(status)
 
         else: